README.md
Welcome to the comprehensive guide designed to help you effortlessly activate your BingX referral code and unlock a world of enhanced trading benefits. For new users, utilizing a referral code like IMOIJ1 is not just an option, it's a gateway to maximizing your initial experience on the BingX platform, offering bonuses up to 7700+ USDT and up to 0.5 discount fees.